Constants.cs: - Updated names of all configuration keys to clearer names. - Updated references and items that use those update constants/keys.
Short.cs:
I wanted to get rid of loose variables like 'url' and 'redirectUrl'.  It's annoying and with enough of them it becomes confusing. It's hard to tell if this is a 'redirectUrl' as a place that is being redirected to, or the redirectUrl as in the url that is used for redirecting users to their link.  Instead all urls will be packaged within a "Link" object.  The tool produces and retrieves Links.  Loose Url variables must die.

A note on that.  I'm not sure if this is an AspNetCore thing, but it appears that objects passed in as parameters to the method with lower case members get matched to their corresponding cases.  Conversly all returns with upper cased members get lower cased when returned as json to the browser.  This means that all javascript should expect lower case member names and all c# should expect lower case members to be matched to their appropriate case in the  object. woof.

appsettings.json: updated to use the new variable name

App.js: Updated to use the Link's 'url' member instead of 'redirectUrl'.

Form: Updated to use Link nomenclature.